Improves distance detection

+++ controller.h
@@ -8,6 +8,6 @@ void TurnLeft();
 void TurnRight();
 void StopMotors();
 
-float DistanceFromSensor();
+int DistanceFromSensor();
 
 #endif
